moved to EL1, tested SVC with vector table

+++ src/system/mmu.c
@@ -1,7 +1,7 @@
 #include "mmu.h"
 
 inline uint64_t read_sctlr_el1() {
-    uint32_t ret;
+    uint64_t ret;
     asm volatile(
         "mrs %[ret], SCTLR_EL1"
         : [ret] "=r" (ret)
